eeconstruct dao layer, implement interface, and organize dao layer unit test code

+++ codefortest/env.go
@@ -12,6 +12,7 @@ import (
 
 	accountModule "github.com/SmartMeshFoundation/Photon/accounts"
 	"github.com/SmartMeshFoundation/Photon/models"
+	"github.com/SmartMeshFoundation/Photon/models/stormdb"
 	"github.com/SmartMeshFoundation/Photon/network/helper"
 	"github.com/SmartMeshFoundation/Photon/network/rpc/contracts"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
@@ -133,10 +134,20 @@ func GetAccountsByAddress(address common.Address) (account TestAccount, err erro
 	return
 }
 
-// GetDb :
-func GetDb() (model *models.ModelDB, err error) {
-	dbPath := path.Join(os.TempDir(), "testxxxx.db")
-	err = os.Remove(dbPath)
-	err = os.Remove(dbPath + ".lock")
-	return models.OpenDb(dbPath)
+// NewTestDB :
+func NewTestDB(dbPath string) models.Dao {
+	if dbPath == "" {
+		dbPath = path.Join(os.TempDir(), "testxxxx.db")
+		err := os.Remove(dbPath)
+		err = os.Remove(dbPath + ".lock")
+		if err != nil {
+			fmt.Println(err)
+		}
+	}
+	var err error
+	db, err := stormdb.OpenDb(dbPath)
+	if err != nil {
+		panic(err)
+	}
+	return db
 }
